Almot done except Python

+++ Controls/MotorController.h
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
+#include <time.h>
 
 /**
  * @brief   Transmits motor control msg to CAN.
@@ -8,4 +9,4 @@
  * @param   CAN_MESSAGE : 64-bits - 63-32 = var1, 31-0 = var2
  * @return  None
  */
-void Motor_Control(uint16_t CAN_ID, uint64_t CAN_MESSAGE);
+void Motor_Control(uint16_t CAN_ID, unsigned long long int CAN_MESSAGE);
